Split master/worker files and make master init db before forking.
